test-requirements.txt
boto==2.38.0
Django==1.4.19
django-appconf==0.4
django-celery==3.1.16
django-celery-email==1.0.4
django-compressor==1.2
django-dbsettings==0.7
django-extensions==1.0.2
django-formwizard==1.0
django-haystack==2.2.0
django-model-utils==2.2
django-mptt==0.6.0
django-notification==0.1.5
django-picklefield==0.1.9
django-positions==0.5.1
django-s3direct==0.3.15
django-tastypie==0.9.10
djangorestframework==2.3.12
celery==3.1.16
amqp==1.4.6
python-dateutil==2.1
python-digest==1.7
python-ldap==2.4.10
python-memcached==1.48
psycopg2==2.5.1
oauth2client==1.2
ordereddict==1.1
orderedset==1.1.2
ply==3.4
google-api-python-client==1.2
elasticsearch==1.2.0
elasticsearch-dsl==0.0.2
elasticutils==0.10
certifi==2015.4.28
M2Crypto==0.22.3
South==0.7.6
unittest2==1.1.0
mock==2.0.0
httplib2==0.9.2
requests==2.13.0